Revert "Add event registration form UI"
